Refactoring where enums are specified and make it possible to get a CommandClass enum
