+++ src/value_classes/value_id.rs
@@ -7,6 +7,103 @@ use std::fmt;
 
 pub use ffi::value_classes::value_id::{ValueGenre, ValueType};
 
+// Mapping comes from https://github.com/OpenZWave/open-zwave-control-panel/blob/master/zwavelib.cpp
+c_like_enum! {
+    CommandClass {
+        NoOperation = 0,
+        Basic = 0x20,
+        ControllerReplication = 0x21,
+        ApplicationStatus = 0x22,
+        ZipServices = 0x23,
+        ZipServer = 0x24,
+        SwitchBinary = 0x25,
+        SwitchMultilevel = 0x26,
+        SwitchAll = 0x27,
+        SwitchToggleBinary = 0x28,
+        SwitchToggleMultilevel = 0x29,
+        ChimneyFan = 0x2A,
+        SceneActivation = 0x2B,
+        SceneActuatorConf = 0x2C,
+        SceneControllerConf = 0x2D,
+        ZipClient = 0x2E,
+        ZipAdvServices = 0x2F,
+        SensorBinary = 0x30,
+        SensorMultilevel = 0x31,
+        Meter = 0x32,
+        Color = 0x33,
+        ZipAdvClient = 0x34,
+        MeterPulse = 0x35,
+        ThermostatHeating = 0x38,
+        ThermostatMode = 0x40,
+        ThermostatOperatingState = 0x42,
+        ThermostatSetpoint = 0x43,
+        ThermostatFanMode = 0x44,
+        ThermostatFanState = 0x45,
+        ClimateControlSchedule = 0x46,
+        ThermostatSetback = 0x47,
+        DoorLockLogging = 0x4C,
+        ScheduleEntryLock = 0x4E,
+        BasicWindowCovering = 0x50,
+        MtpWindowCovering = 0x51,
+        Crc16Encap = 0x56,
+        DeviceResetLocally = 0x5A,
+        CentralScene = 0x5B,
+        ZWavePlusInfo = 0x5E,
+        MultiInstance = 0x60,
+        DoorLock = 0x62,
+        UserCode = 0x63,
+        Configuration = 0x70,
+        Alarm = 0x71,
+        ManufacturerSpecific = 0x72,
+        Powerlevel = 0x73,
+        Protection = 0x75,
+        Lock = 0x76,
+        NodeNaming = 0x77,
+        FirmwareUpdateMd = 0x7A,
+        GroupingNane = 0x7B,
+        RemoteAssociationActivate = 0x7C,
+        RemoteAssociation = 0x7D,
+        Battery = 0x80,
+        Clock = 0x81,
+        Hail = 0x82,
+        WakeUp = 0x84,
+        Association = 0x85,
+        Version = 0x86,
+        Indicator = 0x87,
+        Proprietary = 0x88,
+        Language = 0x89,
+        Time = 0x8A,
+        TimeParameters = 0x8B,
+        GeographicLocation = 0x8C,
+        Composite = 0x8D,
+        MultiInstanceAssociation = 0x8E,
+        MultiCmd = 0x8F,
+        EnergyProduction = 0x90,
+        ManufacturerProprietary = 0x91,
+        ScreenMd = 0x92,
+        ScreenAttributes = 0x93,
+        SimpleAvControl = 0x94,
+        AvContentDirectoryMd = 0x95,
+        AvRendererStatus = 0x96,
+        AvContentSearchMd = 0x97,
+        Security = 0x98,
+        AvTaggingMd = 0x99,
+        IpConfiguration = 0x9A,
+        AssociationCommandConfiguration = 0x9B,
+        SensorAlarm = 0x9C,
+        SilenceAlarm = 0x9D,
+        SensorConfiguration = 0x9E,
+        Mark = 0x9F,
+        NonInteroperable = 0xF0
+    }
+}
+
+impl fmt::Display for CommandClass {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        fmt::Debug::fmt(self, f)
+    }
+}
+
 use ffi::utils::{
     rust_string_creator, rust_vec_creator, rust_string_vec_creator,
     recover_string, recover_vec
@@ -137,6 +234,10 @@ impl ValueID {
 
     pub fn get_command_class_id(&self) -> u8 {
         unsafe { extern_value_id::value_id_get_command_class_id(&self.as_ozw_vid()) }
+    }
+
+    pub fn get_command_class(&self) -> Option<CommandClass> {
+        CommandClass::from_u8(self.get_command_class_id())
     }
 
     pub fn get_instance(&self) -> u8 {
@@ -375,7 +476,7 @@ impl fmt::Display for ValueID {
                        self.get_id(),
                        self.get_node_id(),
                        node_name,
-                       self.get_command_class_id(),
+                       self.get_command_class().map_or(String::from("???"), |cc| cc.to_string()),
                        self.get_type(),
                        self.get_label(),
                        self.as_string().unwrap_or(String::from("???"))))
@@ -384,7 +485,7 @@ impl fmt::Display for ValueID {
 
 impl fmt::Debug for ValueID {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "ValueID {{ home_id: {:?}, node_id: {:?}, genre: {:?}, command_class_id: {:?}, \
+        write!(f, "ValueID {{ home_id: {:?}, node_id: {:?}, genre: {:?}, command_class: {:?}, \
                    instance: {:?}, index: {:?}, type: {:?}, id: {:?}, \
                    label: {:?}, units: {:?}, help: {:?}, min: {:?}, max: {:?}, is_read_only: {:?}, \
                    is_write_only: {:?}, is_set: {:?}, is_polled: {:?}, \
@@ -395,7 +496,7 @@ impl fmt::Debug for ValueID {
                self.get_home_id(),
                self.get_node_id(),
                self.get_genre(),
-               self.get_command_class_id(),
+               self.get_command_class(),
                self.get_instance(),
                self.get_index(),
                self.get_type(),
